move all typings, type guards, etc relative to DB model data to corresponding model definition file
